moved tests to a subfolder
