+++ examples/typeorm/pg/test/connect.ts
@@ -13,15 +13,15 @@
 // limitations under the License.
 
 import t from 'tap';
-import {connect} from './connect.mjs';
+import {connect} from '../connect';
 
-t.test('mysql2 typeorm mjs', async t => {
-  const { dataSource, close } = await connect({
-    instanceConnectionName: process.env.MYSQL_IAM_CONNECTION_NAME,
-    username: process.env.MYSQL_IAM_USER,
-    database: process.env.MYSQL_DB,
+t.test('pg typeorm ts', async t => {
+  const {dataSource, close} = await connect({
+    instanceConnectionName: process.env.POSTGRES_IAM_CONNECTION_NAME,
+    username: process.env.POSTGRES_IAM_USER,
+    database: process.env.POSTGRES_DB,
   });
-  const [{ now }] = await dataSource.manager.query('SELECT NOW() as now');
+  const [{now}] = await dataSource.manager.query('SELECT NOW() as now');
   t.ok(now.getTime(), 'should have valid returned date object');
   await close();
 });
