+++ examples/typeorm/tedious/test/connect.cjs
@@ -12,16 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import t from 'tap';
-import {connect} from './connect.mjs';
+const t = require('tap');
+const connector = require('../connect.cjs');
 
-t.test('pg typeorm mjs', async t => {
-  const { dataSource, close } = await connect({
-    instanceConnectionName: process.env.POSTGRES_IAM_CONNECTION_NAME,
-    username: process.env.POSTGRES_IAM_USER,
-    database: process.env.POSTGRES_DB,
+t.test('tedious typeorm cjs', async t => {
+  const { dataSource, close } = await connector.connect({
+    instanceConnectionName: process.env.SQLSERVER_CONNECTION_NAME,
+    username: process.env.SQLSERVER_USER,
+    password: process.env.SQLSERVER_PASS,
+    database: process.env.SQLSERVER_DB,
   });
-  const [{ now }] = await dataSource.manager.query('SELECT NOW() as now');
+  const [{ now }] = await dataSource.manager.query('SELECT GETUTCDATE() as now')
   t.ok(now.getTime(), 'should have valid returned date object');
   await close();
 });
